Delete unuseable file and fix bugs in my code
