+++ src/index.js
@@ -1,16 +1,27 @@
 import './style.css';
-import displayScores from './modules/DOM.js';
-import postScores from './modules/api-Call.js';
- 
+import {getScores, postScores} from './modules/api-Call.js';
 
+const displayScores = () => {
+    getScores().then((scores) => {
+        const newScores = document.querySelector('.new-scores');
+        newScores.innerHTML = '';
+  scores.result.forEach((el) => {
+    const li = document.createElement('li');
+    li.textContent =`${el.user}: ${el.score}`;
+    newScores.appendChild(li);
+    console.log(newScores);
+  });
+  return scores;
+    })
+  };
+  
 const addUserScore = document.querySelector('.submit');
 const refresh = document.querySelector('.refresh');
 
 addUserScore.addEventListener('click', (e) => {
     e.preventDefault();
-    
     const nameInput = document.querySelector('.name-input').value;
-    const scoreInput = +document.querySelector('.score-input').value;
+    const scoreInput = document.querySelector('.score-input').value;
     if (nameInput && scoreInput) {
       postScores(nameInput, scoreInput);
      displayScores();
@@ -18,6 +29,8 @@ addUserScore.addEventListener('click', (e) => {
      = document.querySelector('.score-input').value ='';
     }
   });
+
+
 
   refresh.addEventListener('click',(e) => {
       displayScores();
